Fixes metadata race conditions Replaces empty structs by nil Removes unused functions and structs
