+++ lib/backend/iaas/stacks/gcp/stack.go
@@ -47,7 +47,7 @@ type stack struct {
 
 // NullStack is not exposed through API, is needed essentially by tests
 func NullStack() *stack { // nolint
-	return &stack{}
+	return nil
 }
 
 func (s *stack) IsNull() bool {
