+++ lib/backend/resources/abstract/enum.go
@@ -15,4 +15,5 @@ const (
 	SecurityGroupResource
 	SubnetResource
 	VolumeResource
+	ShareResource
 )
