+++ lib/utils/retry/errors.go
@@ -20,8 +20,6 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/CS-SI/SafeScale/v22/lib/utils/data"
-	"github.com/CS-SI/SafeScale/v22/lib/utils/debug/callstack"
 	"github.com/CS-SI/SafeScale/v22/lib/utils/fail"
 	"github.com/CS-SI/SafeScale/v22/lib/utils/strprocess"
 	"github.com/CS-SI/SafeScale/v22/lib/utils/temporal"
@@ -31,26 +29,12 @@ import (
 type ErrTimeout = fail.ErrTimeout
 
 // TimeoutError creates an error of type ErrTimeout
-func TimeoutError(err error, limit time.Duration, actual time.Duration, options ...data.ImmutableKeyValue) fail.Error {
+func TimeoutError(err error, limit time.Duration, actual time.Duration) fail.Error {
 	var (
-		msg      string
-		decorate bool
+		msg string
 	)
 
-	if len(options) > 0 {
-		for _, v := range options {
-			switch v.Key() { // nolint
-			case "callstack":
-				// no panics if value is not a bool
-				decorate, _ = v.Value().(bool) // nolint
-			}
-		}
-	}
-
 	msg = fmt.Sprintf("retries timed out after %s (timeout defined at %s)", temporal.FormatDuration(actual), temporal.FormatDuration(limit))
-	if decorate {
-		msg = callstack.DecorateWith(msg, "", "", 0)
-	}
 	return fail.TimeoutError(err, limit, msg)
 }
